ibmcloud: isolated auxillary functions and methods
Moved helper functions and methods into different  file

Fixes: #1161
